move flag constants to dedicated file
