+++ services/QuillLMS/client/app/bundles/Shared/components/shared/flagDropdown.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { flagOptions } from '../../../../constants/evidence'
+import { flagOptions } from '../../../../constants/flagOptions'
 
 const flagOptionList = () => {
   return ( flagOptions.map((flagOption) => (
