+++ services/QuillLMS/client/app/bundles/Teacher/components/assignment_flow/create_unit/custom_activity_pack/flag_filters.tsx
@@ -3,7 +3,7 @@ import _  from 'lodash';
 
 import { Activity } from './interfaces'
 import { FLAG_FILTERS } from './shared'
-import { flagOptions } from '../../../../../../constants/evidence';
+import { flagOptions } from '../../../../../../constants/flagOptions';
 
 const smallWhiteCheckSrc = `${process.env.CDN_URL}/images/shared/check-small-white.svg`
