+++ services/QuillLMS/client/app/constants/evidence.ts
@@ -43,44 +43,6 @@ export const TEXT = 'text';
 export const FIRST_STRONG_EXAMPLE = 'first_strong_example';
 export const SECOND_STRONG_EXAMPLE = 'second_strong_example';
 
-export const flagOptions = [
-  {
-    label: 'alpha',
-    value: 'alpha'
-  },
-  {
-    label: 'private',
-    value: 'private'
-  },
-  {
-    label: 'beta',
-    value: 'beta'
-  },
-  {
-    label: 'gamma',
-    value: 'gamma'
-  },
-  {
-    label: 'Evidence Beta 1',
-    value: 'evidence_beta1'
-  },
-  {
-    label: 'Evidence Beta 2',
-    value: 'evidence_beta2'
-  },
-  {
-    label: 'College Board',
-    value: 'college_board'
-  },
-  {
-    label: 'production',
-    value: 'production'
-  },
-  {
-    label: 'archived',
-    value: 'archived'
-  },
-];
 
 export const courseOptions = [
   {
