+++ services/QuillLMS/client/app/constants/flagOptions.ts
@@ -1,0 +1,38 @@
+export const flagOptions = [
+  {
+    label: 'alpha',
+    value: 'alpha'
+  },
+  {
+    label: 'private',
+    value: 'private'
+  },
+  {
+    label: 'beta',
+    value: 'beta'
+  },
+  {
+    label: 'gamma',
+    value: 'gamma'
+  },
+  {
+    label: 'Evidence Beta 1',
+    value: 'evidence_beta1'
+  },
+  {
+    label: 'Evidence Beta 2',
+    value: 'evidence_beta2'
+  },
+  {
+    label: 'College Board',
+    value: 'college_board'
+  },
+  {
+    label: 'production',
+    value: 'production'
+  },
+  {
+    label: 'archived',
+    value: 'archived'
+  },
+];
